Fix for encoding issue with one or more zero high-order bytes in an mpi
The function mbedtls_mpi_write_binary will always fill in the entire buffer, providing leading zeros as necessary.  Instead of providing an mpi-specific length to this call, and because we know we always want a fixed-size buffer for any mpi, call this function with a fixed length of RIOT_COORDMAX.

Also take this opportunity to limit the chance for similar issues elsewhere (transitioning to RIOT_COORDMAX when we know we want a buffer of that size) and don't rely on the correctness of encBufferLen.